Refactor index.ts, extract Formsy class to a separate file
